refactor: link-popup is now a vue component
